requirements.txt
djangorestframework
pytest-django
factory-boy
psycopg2-binary